# Conflicts:
#	src/main/java/game/service/BulletDataProcessor.java

+++ src/main/java/game/datatype/shield/ShieldParent.java
@@ -6,22 +6,22 @@ import game.interfaces.Shield;
 
 public class ShieldParent extends ItemParent implements Shield {
 
-    private long protection;
-    private long maxProtectionValue;
+    private double protection;
+    private double maxProtectionValue;
 
-    public long getProtection() {
+    public double getProtection() {
         return protection;
     }
 
-    public void setProtection(long protection) {
+    public void setProtection(double protection) {
         this.protection = protection;
     }
 
-    public long getMaxProtectionValue() {
+    public double getMaxProtectionValue() {
         return maxProtectionValue;
     }
 
-    public void setMaxProtectionValue(long maxProtectionValue) {
+    public void setMaxProtectionValue(double maxProtectionValue) {
         this.maxProtectionValue = maxProtectionValue;
     }
 
@@ -35,7 +35,7 @@ public class ShieldParent extends ItemParent implements Shield {
     }
 
     @Override
-    public void decreaseProtection(long value) {
+    public void decreaseProtection(double value) {
         this.protection -= value;
         if (this.protection < 0L)
             this.protection = 0L;
@@ -49,7 +49,7 @@ public class ShieldParent extends ItemParent implements Shield {
     }
 
     @Override
-    public void increaseShieldPowerBy(long amount) {
+    public void increaseShieldPowerBy(double amount) {
         protection += amount;
         if (protection > maxProtectionValue) {
             protection = maxProtectionValue;
