+++ src/main/test/game/datatype/shield/ShieldParentTest.java
@@ -35,27 +35,27 @@ public class ShieldParentTest {
     public void testShouldDecreaseShieldProtection_AS() {
         as.applyEffect(player);
 
-        player.getShield().decreaseProtection(10L);
+        player.getShield().decreaseProtection(10.0);
 
-        Assert.assertEquals(player.getShield().getProtection(), ShieldConfig.ATOM_SHIELD_PROTECTION - 10L);
+        Assert.assertEquals((long)player.getShield().getProtection(), (long)ShieldConfig.ATOM_SHIELD_PROTECTION - 10L);
     }
 
     @Test
     public void testShouldDecreaseShieldProtection_PS() {
         ps.applyEffect(player);
 
-        player.getShield().decreaseProtection(10L);
+        player.getShield().decreaseProtection(10.0);
 
-        Assert.assertEquals(player.getShield().getProtection(), ShieldConfig.PLASMA_SHIELD_PROTECTION - 10L);
+        Assert.assertEquals((long)player.getShield().getProtection(), (long)ShieldConfig.PLASMA_SHIELD_PROTECTION - 10L);
     }
 
     @Test
     public void testShouldDecreaseShieldProtection_NS() {
         ns.applyEffect(player);
 
-        player.getShield().decreaseProtection(10L);
+        player.getShield().decreaseProtection(10.0);
 
-        Assert.assertEquals(player.getShield().getProtection(), ShieldConfig.NORMAL_SHIELD_PROTECTION - 10L);
+        Assert.assertEquals((long)player.getShield().getProtection(), (long)ShieldConfig.NORMAL_SHIELD_PROTECTION - 10L);
     }
 
     @Test
@@ -64,7 +64,7 @@ public class ShieldParentTest {
 
         player.getShield().decreaseProtection(ShieldConfig.ATOM_SHIELD_PROTECTION + 1L);
 
-        Assert.assertEquals(player.getShield().getProtection(), 0L);
+        Assert.assertEquals((long)player.getShield().getProtection(), 0L);
     }
 
     @Test
@@ -73,7 +73,7 @@ public class ShieldParentTest {
 
         player.getShield().decreaseProtection(ShieldConfig.PLASMA_SHIELD_PROTECTION + 1L);
 
-        Assert.assertEquals(player.getShield().getProtection(), 0L);
+        Assert.assertEquals((long)player.getShield().getProtection(), 0L);
     }
 
     @Test
@@ -82,7 +82,7 @@ public class ShieldParentTest {
 
         player.getShield().decreaseProtection(ShieldConfig.NORMAL_SHIELD_PROTECTION + 1L);
 
-        Assert.assertEquals(player.getShield().getProtection(), 0L);
+        Assert.assertEquals((long)player.getShield().getProtection(), 0L);
     }
 
     public void testShouldIncreaseShieldProtection_AS() {
@@ -92,7 +92,7 @@ public class ShieldParentTest {
 
         player.getShield().increaseShieldPower();
 
-        Assert.assertEquals(player.getShield().getProtection(), 1L);
+        Assert.assertEquals((long)player.getShield().getProtection(), 1L);
     }
 
     @Test
@@ -101,7 +101,7 @@ public class ShieldParentTest {
         player.getShield().setProtection(0L);
         player.getShield().increaseShieldPower();
 
-        Assert.assertEquals(player.getShield().getProtection(), 1L);
+        Assert.assertEquals((long)player.getShield().getProtection(), 1L);
     }
 
     @Test
@@ -111,7 +111,7 @@ public class ShieldParentTest {
 
         player.getShield().increaseShieldPower();
 
-        Assert.assertEquals(player.getShield().getProtection(), 1L);
+        Assert.assertEquals((long)player.getShield().getProtection(), 1L);
     }
 
     public void testShouldIncreaseShieldProtection_AS_MAX() {
@@ -121,7 +121,7 @@ public class ShieldParentTest {
 
         player.getShield().increaseShieldPower();
 
-        Assert.assertEquals(player.getShield().getProtection(), ShieldConfig.ATOM_SHIELD_PROTECTION);
+        Assert.assertEquals((long)player.getShield().getProtection(), (long)ShieldConfig.ATOM_SHIELD_PROTECTION);
     }
 
     @Test
@@ -130,7 +130,7 @@ public class ShieldParentTest {
         player.getShield().setProtection(ShieldConfig.PLASMA_SHIELD_PROTECTION);
         player.getShield().increaseShieldPower();
 
-        Assert.assertEquals(player.getShield().getProtection(), ShieldConfig.PLASMA_SHIELD_PROTECTION);
+        Assert.assertEquals((long)player.getShield().getProtection(), (long)ShieldConfig.PLASMA_SHIELD_PROTECTION);
     }
 
     @Test
@@ -140,6 +140,6 @@ public class ShieldParentTest {
 
         player.getShield().increaseShieldPower();
 
-        Assert.assertEquals(player.getShield().getProtection(), ShieldConfig.NORMAL_SHIELD_PROTECTION);
+        Assert.assertEquals((long)player.getShield().getProtection(), (long)ShieldConfig.NORMAL_SHIELD_PROTECTION);
     }
 }
